Removed cast for getAIRange that forgot to remove last commit

+++ st_demo/source/st_demo.cpp
@@ -439,7 +439,7 @@ int stDemo::getDefaultLightSetIndex()
 }
 stRange* stDemo::getAIRange()
 {
-    return (u32) & this->aiRange;
+    return &this->aiRange;
 }
 bool stDemo::isAdventureStage()
 {
